Renamed uwsgi.py in uwsgicache.py to avoid module import naming clash in python2
